Guard condition to make sure timed content postmeta not updated when save_post action happens outside the admin UI (cron job)

+++ plugins/greatermedia-timed-content/inc/class-greatermedia-timed-content.php
@@ -141,31 +141,35 @@ class GreaterMediaTimedContent extends VisualShortcode {
 
 		if ( post_type_supports( $post->post_type, 'timed-content' ) ) {
 
-			$exp_mm = isset( $_POST['hidden_exp_mm'] ) ? intval( $_POST['hidden_exp_mm'] ) : '0';
-			$exp_jj = isset( $_POST['hidden_exp_jj'] ) ? intval( $_POST['hidden_exp_jj'] ) : '0';
-			$exp_aa = isset( $_POST['hidden_exp_aa'] ) ? intval( $_POST['hidden_exp_aa'] ) : '0';
-			$exp_hh = isset( $_POST['hidden_exp_hh'] ) ? intval( $_POST['hidden_exp_hh'] ) : '0';
-			$exp_mn = isset( $_POST['hidden_exp_mn'] ) ? str_pad( intval( $_POST['hidden_exp_mn'] ), 2, '0', STR_PAD_LEFT ) : '00';
+			if ( isset( $_SERVER['REQUEST_METHOD'] ) && 'POST' === $_SERVER['REQUEST_METHOD'] && isset( $_POST['hidden_exp_mm'] ) ) {
 
-			$exp_str = "{$exp_mm} {$exp_jj} {$exp_aa} {$exp_hh} {$exp_mn}";
-			if ( '0 0 0 0 00' !== trim( $exp_str ) ) {
-				$exp_date      = DateTime::createFromFormat( 'n j Y G i', $exp_str );
-				$exp_timestamp = $exp_date->getTimestamp();
-			} else {
-				$exp_timestamp = '';
-			}
+				$exp_mm = isset( $_POST['hidden_exp_mm'] ) ? intval( $_POST['hidden_exp_mm'] ) : '0';
+				$exp_jj = isset( $_POST['hidden_exp_jj'] ) ? intval( $_POST['hidden_exp_jj'] ) : '0';
+				$exp_aa = isset( $_POST['hidden_exp_aa'] ) ? intval( $_POST['hidden_exp_aa'] ) : '0';
+				$exp_hh = isset( $_POST['hidden_exp_hh'] ) ? intval( $_POST['hidden_exp_hh'] ) : '0';
+				$exp_mn = isset( $_POST['hidden_exp_mn'] ) ? str_pad( intval( $_POST['hidden_exp_mn'] ), 2, '0', STR_PAD_LEFT ) : '00';
 
-			$local_to_gmt_time_offset = get_option( 'gmt_offset' ) * - 1 * 3600;
-			$exp_timestamp_gmt        = $exp_timestamp + $local_to_gmt_time_offset;
-			delete_post_meta( $post_id, 'post_expiration' );
-			if ( '' !== $exp_timestamp ) {
-				add_post_meta( $post_id, 'post_expiration', $exp_timestamp );
-			}
+				$exp_str = "{$exp_mm} {$exp_jj} {$exp_aa} {$exp_hh} {$exp_mn}";
+				if ( '0 0 0 0 00' !== trim( $exp_str ) ) {
+					$exp_date      = DateTime::createFromFormat( 'n j Y G i', $exp_str );
+					$exp_timestamp = $exp_date->getTimestamp();
+				} else {
+					$exp_timestamp = '';
+				}
 
-			// If the expiration date is in the future, set a cron to expire the post
-			if ( $exp_timestamp_gmt > gmdate( 'U' ) ) {
-				wp_clear_scheduled_hook( 'greatermedia_expire_post', array( $post_id ) ); // clear anything else in the system
-				wp_schedule_single_event( $exp_timestamp_gmt, 'greatermedia_expire_post', array( $post_id ) );
+				$local_to_gmt_time_offset = get_option( 'gmt_offset' ) * - 1 * 3600;
+				$exp_timestamp_gmt        = $exp_timestamp + $local_to_gmt_time_offset;
+				delete_post_meta( $post_id, 'post_expiration' );
+				if ( '' !== $exp_timestamp ) {
+					add_post_meta( $post_id, 'post_expiration', $exp_timestamp );
+				}
+
+				// If the expiration date is in the future, set a cron to expire the post
+				if ( $exp_timestamp_gmt > gmdate( 'U' ) ) {
+					wp_clear_scheduled_hook( 'greatermedia_expire_post', array( $post_id ) ); // clear anything else in the system
+					wp_schedule_single_event( $exp_timestamp_gmt, 'greatermedia_expire_post', array( $post_id ) );
+				}
+
 			}
 
 		} else {
